requirements/base.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-countries==7.4.2
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-rest-swagger==2.2.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.57
Pillow==9.2.0
pyparsing==3.0.9
pytz==2022.4
pyotp==2.8.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
argon2-cffi==21.3.0
redis==4.3.4
flower==1.2.0
django-celery-email==3.0.0
django-tenants==3.4.5
django-enumchoicefield==3.0.1
PyJWT==2.5.0
django-templated-mail==1.1.1
django-extensions==3.2.1
django-storages==1.13.1
boto3==1.24.96
Werkzeug==2.2.2
numpy==1.23.4
pandas==1.5.1
django-safedelete==1.3.1
social-auth-app-django==5.0.0
django-authlib==0.13.1
Authlib==1.2.0
jsonschema==4.17.3
celery[sqs]==5.2.7
openpyxl==3.0.7
multiprocesspandas==1.1.5
djangorestframework-jwt
django-permissions-policy==4.18.0
django-csp==3.7
Unidecode==1.3.7
matplotlib==3.7.3
python-pptx==0.6.23
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
msal
channels
django-cron